+++ packages/entities/entities-plugins/src/components/free-form/shared/const.ts
@@ -14,11 +14,3 @@ export const partialEndpoints = {
     getAll: '/{workspace}/partials',
   },
 }
-
-// Redis configuration fields order
-export const fieldsOrder = [
-  'host', 'port', 'connection_is_proxied', 'database', 'username', 'password',
-  'sentinel_master', 'sentinel_role', 'sentinel_nodes', 'sentinel_username', 'sentinel_password',
-  'cluster_nodes', 'cluster_max_redirections', 'ssl', 'ssl_verify', 'server_name',
-  'keepalive_backlog', 'keepalive_pool_size', 'read_timeout', 'send_timeout', 'connect_timeout',
-]