+++ packages/entities/entities-redis-configurations/sandbox/pages/RedisConfigurationFormPage.vue
@@ -46,6 +46,7 @@ const kongManagerConfig: KongManagerRedisConfigurationFormConfig = {
   workspace: 'default',
   apiBaseUrl: '/kong-manager', // For local dev server proxy
   cloudAuthAvailable: true,
+  isPortReferenceable: true,
   cancelRoute: { name: 'redis-configuration-list' },
 }