+++ packages/entities/entities-redis-configurations/src/components/RedisConfigurationForm.vue
@@ -231,7 +231,15 @@
             tooltipAttributes: { maxWidth: '400' },
           }"
           :readonly="form.readonly"
-          type="number"
+        />
+        <VaultSecretPickerProvider
+          v-if="redisType === RedisType.HOST_PORT_EE && config.isPortReferenceable"
+          class="secret-picker-provider"
+          data-testid="secret-picker-provider-for-port"
+          :disabled="form.readonly"
+          :update="v => form.fields.config.port = v"
+          :value="String(form.fields.config.port ?? '')"
+          @open="(value, update) => setUpVaultSecretPicker(value, update)"
         />
 
         <KCheckbox