+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.26](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.25...@kong-ui-public/analytics-geo-map@1.4.26) (2025-12-04)
+
+
+### Bug Fixes
+
+* **deps:** update non-major [@kong](https://github.com/kong) dependencies with stable versions ([#2706](https://github.com/Kong/public-ui-components/issues/2706)) ([2e83580](https://github.com/Kong/public-ui-components/commit/2e835806501a95a5e36e35c36f9bc60aeb678329))
+
+
+
+
+
 ## [1.4.25](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.24...@kong-ui-public/analytics-geo-map@1.4.25) (2025-12-03)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-geo-map